+++ frontend_web/src/api/chat/types.ts
@@ -1,5 +1,5 @@
-import type { MastraMessageV2 } from '@mastra/core';
 import type { Message } from '@ag-ui/core';
+import { ContentPart, ToolInvocation } from '@/types/message.ts';
 
 // TODO: This awkwardly shoves AGUI message list into the Mastra format.
 // I didn't want to completely rewrite the UI, still figuring out where to go here.
@@ -13,7 +13,31 @@ export interface ChatListItem {
   metadata?: Record<string, unknown>;
 }
 
-export type MessageResponse = MastraMessageV2;
+type JSONValue =
+  | null
+  | string
+  | number
+  | boolean
+  | {
+      [value: string]: JSONValue;
+    }
+  | Array<JSONValue>;
+
+export type MessageResponse = {
+  id: string;
+  role: 'user' | 'assistant' | 'system';
+  createdAt: Date;
+  threadId?: string;
+  resourceId?: string;
+  type?: string;
+  format: 2;
+  parts: ContentPart[];
+  content?: string;
+  toolInvocations?: ToolInvocation[];
+  reasoning?: string;
+  annotations?: JSONValue[] | undefined;
+  metadata?: Record<string, unknown>;
+};
 
 export type MessagePage = {
   messages: MessageResponse[];